requirements.txt
Unidecode==1.0.23
python_dateutil==2.8.0
PyYAML==5.1
